+++ src/main.rs
@@ -15,5 +15,6 @@ fn main() -> ! {
         Picodata::Expel(args) => cli::expel::main(args),
         Picodata::Connect(args) => cli::connect::main(args),
         Picodata::Admin(args) => cli::admin::main(args),
+        Picodata::Config(cli::args::Config::Default(args)) => cli::default_config::main(args),
     }
 }